+++ src/lib/components/table/evm-transactions/EvmTransactionsTableMobileCard.tsx
@@ -5,7 +5,7 @@ import { MobileCardTemplate } from "../MobileCardTemplate";
 import { MobileLabel } from "../MobileLabel";
 import { useInternalNavigate } from "lib/app-provider";
 import { EvmMethodChip } from "lib/components/EvmMethodChip";
-import { EvmToCell } from "lib/components/EvmToCell";
+import { EvmToCell } from "lib/components/evm-to-cell";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
 import type { TxDataWithTimeStampJsonRpc } from "lib/services/types";