+++ src/lib/pages/evm-tx-details/components/evm-tx-method/EvmTxCreateContract.tsx
@@ -1,4 +1,4 @@
-import { Flex, Text } from "@chakra-ui/react";
+import { Flex, Tag, Text } from "@chakra-ui/react";
 
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
@@ -6,6 +6,7 @@ import type { TxDataJsonRpc } from "lib/services/types";
 
 import { EvmInfoLabelValue } from "./EvmInfoLabelValue";
 import { EvmTxMethodAccordion } from "./EvmTxMethodAccordion";
+import { useCreatedContractsByEvmTxHash } from "lib/services/tx";
 
 interface EvmTxCreateContractProps {
   evmTxData: TxDataJsonRpc;
@@ -16,8 +17,10 @@ export const EvmTxCreateContract = ({
 }: EvmTxCreateContractProps) => {
   const { from } = evmTxData.tx;
   const { contractAddress } = evmTxData.txReceipt;
-
-  // TODO: fix contract addresses
+  const { data: contracts = [] } = useCreatedContractsByEvmTxHash(
+    evmTxData.tx.hash,
+    contractAddress
+  );
 
   return (
     <EvmTxMethodAccordion
@@ -25,19 +28,28 @@ export const EvmTxCreateContract = ({
       content={
         <Flex gap={1}>
           Create{" "}
-          {contractAddress ? (
-            <Flex gap={1} align="center">
-              <CustomIcon
-                name="contract-address"
-                boxSize={3}
-                color="primary.main"
-              />
-              <ExplorerLink
-                value={contractAddress}
-                type="evm_contract_address"
-                showCopyOnHover
-              />
-            </Flex>
+          {contracts.length > 0 ? (
+            contracts.length > 1 ? (
+              <>
+                <Tag variant="primary" px={3}>
+                  {contracts.length}
+                </Tag>
+                <Text>contracts</Text>
+              </>
+            ) : (
+              <Flex gap={1} align="center">
+                <CustomIcon
+                  name="contract-address"
+                  boxSize={3}
+                  color="primary.main"
+                />
+                <ExplorerLink
+                  value={contracts[0]}
+                  type="evm_contract_address"
+                  showCopyOnHover
+                />
+              </Flex>
+            )
           ) : (
             <Text variant="body2" color="text.disabled">
               -
@@ -61,20 +73,24 @@ export const EvmTxCreateContract = ({
       <EvmInfoLabelValue
         label="Created Contract"
         value={
-          contractAddress ? (
-            <Flex gap={1} align="center">
-              <CustomIcon
-                name="contract-address"
-                boxSize={3}
-                color="primary.main"
-              />
-              <ExplorerLink
-                value={contractAddress}
-                type="evm_contract_address"
-                showCopyOnHover
-                textFormat="normal"
-                fixedHeight={false}
-              />
+          contracts.length > 0 ? (
+            <Flex direction="column" gap={2}>
+              {contracts.map((contract) => (
+                <Flex key={contract} gap={1} align="center">
+                  <CustomIcon
+                    name="contract-address"
+                    boxSize={3}
+                    color="primary.main"
+                  />
+                  <ExplorerLink
+                    value={contract}
+                    type="evm_contract_address"
+                    showCopyOnHover
+                    textFormat="normal"
+                    fixedHeight={false}
+                  />
+                </Flex>
+              ))}
             </Flex>
           ) : (
             <Text variant="body2" color="text.disabled">